% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhddmjagsutils.R
\name{diagnostic}
\alias{diagnostic}
\title{give diagnostic from MCMC output}
\usage{
diagnostic(object, params = "all", exclude = NULL, ISB = TRUE, exact = TRUE)
}
\arguments{
\item{object}{rjags object}

\item{exact}{}
}
\value{

}
\description{
give diagnostic from MCMC output
}
