% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhddmjagsutils.R
\name{jellyfish}
\alias{jellyfish}
\title{Plots posterior distributions in a jellyfish}
\usage{
jellyfish(samples, parameter, reorder = FALSE, filename = NULL)
}
\arguments{
\item{samples}{rjags object}

\item{parameter}{a string with the parameter of interest}

\item{reorder}{order the distributions? defaults False}

\item{filename}{optional file location to save the plot}
}
\value{

}
\description{
Plots posterior distributions in a jellyfish
}
