% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhddmjagsutils.R
\name{MCMCoutput}
\alias{MCMCoutput}
\title{summarize MCMC output}
\usage{
MCMCoutput(object, params = "all", exclude = NULL, ISB = TRUE, exact = TRUE)
}
\arguments{
\item{object}{rjags object}

\item{params}{name of parameter}

\item{exact}{}
}
\value{

}
\description{
summarize MCMC output
}
