% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn17.R
\docType{data}
\name{mn17}
\alias{mn17}
\title{EEG decision-making data from Nunez et al. (2017)}
\format{
A data frame with 9180 observations of 10 variables.
\describe{
\item{subject}{indicating the subject number.}
\item{rt}{The response time, in milliseconds.}
\item{jitter}{amount of temporal jitter introduced into the stimuli, in milliseconds.}
\item{noise}{amount of visual noise added to the stimuli, as a percentage of maximum contrast.}
\item{n200trialerp_c1r}{N200 ERP component amplitude for responses.}
\item{n200trialerplat_c1r}{N200 ERP component latency for responses.}
\item{p200trialerp_c1n}{P200 ERP component amplitude for responses}
\item{p200trialerplat_c1n}{P200 ERP component latency for responses.}
\item{goodtrials}{indicating which trials to exclude}
}
}
\source{
\url{https://github.com/mdnunez/mcntoolbox}
}
\usage{
mn17
}
\description{
This dataset contains EEG data from a decision-making task described in Nunez et al. (2017)
}
\details{
Nunez, M. D., Vandekerckhove, J., & Srinivasan, R. (2017).
How attention influences perceptual decision making: Single-trial EEG correlates of drift-diffusion model parameters.
Journal of Mathematical Psychology, 76, 117-130.
}
\keyword{datasets}
