% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhddmjagsutils.R
\name{recovery}
\alias{recovery}
\title{Plot parameter recovery}
\usage{
recovery(samples, truevals, filename = NULL)
}
\arguments{
\item{samples}{samples the rjags object}

\item{truevals}{List of true parameter values (the genparam list)}

\item{filename}{optional location to save plot}
}
\value{
ggplot of parameter recovery
}
\description{
Plot parameter recovery
}
