% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhddmjagsutils.R
\name{rsquared_pred}
\alias{rsquared_pred}
\title{RSQUARED_PRED}
\usage{
rsquared_pred(trueval, predval)
}
\arguments{
\item{trueval}{a numeric vector}

\item{predval}{a numeric vector}
}
\value{
a numeric
}
\description{
Calculates R^2_prediction for data and statistics derived from data
}
