% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhddmjagsutils.R
\name{simulratcliff}
\alias{simulratcliff}
\title{Simulate diffusion models faster}
\usage{
simulratcliff(
  N = 100,
  Alpha = 1,
  Tau = 0.4,
  Nu = 1,
  Beta = 0.5,
  rangeTau = 0,
  rangeBeta = 0,
  Eta = 0.3,
  Varsigma = 1
)
}
\arguments{
\item{Varsigma}{}
}
\value{
Vector with reaction times (in seconds) multiplied by the response vector such that negative reaction times encode response B and positive reaction times encode response A
}
\description{
Generates data according to a drift diffusion model with optional trial-to-trial variability faster.
Converted from simuldiff.m MATLAB script by Joachim Vandekerckhove,
Then converted from pyhddmjags utils python script by Kianté Fernandez
See also http://ppw.kuleuven.be/okp/dmatoolbox.
}
